services/TS29222_CAPIF_Logging_API_Invocation_API/requirements.txt
connexion[swagger-ui] >= 2.6.0; python_version>="3.6"
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
setuptools >= 21.0.0
Flask == 2.0.3
pymongo == 4.0.1
elasticsearch == 8.4.3
flask_jwt_extended == 4.4.4
opentelemetry-instrumentation == 0.38b0
opentelemetry-instrumentation-flask == 0.38b0
opentelemetry-instrumentation-redis == 0.38b0
opentelemetry-instrumentation-pymongo == 0.38b0
opentelemetry-exporter-otlp == 1.17.0
opentelemetry-exporter-jaeger==1.17.0
fluent == 0.10.0
fluent-logger == 0.10.0
opentelemetry-api == 1.17.0
opentelemetry-sdk == 1.17.0
flask_executor == 1.0.0
pyopenssl == 23.0.0
